Move ServiceManager and ServiceContext to engine
