+++ src/NUnitEngine/nunit.engine.tests/Services/Fakes/FakeService.cs
@@ -1,8 +1,13 @@
 ﻿// Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
+using System;
+
 namespace NUnit.Engine.Services.Tests.Fakes
 {
-    public class FakeService : IService
+    interface IFakeService
+    { }
+
+    public class FakeService : IFakeService, IService
     {
         IServiceLocator IService.ServiceContext { get; set; }
 
@@ -22,10 +27,16 @@ namespace NUnit.Engine.Services.Tests.Fakes
         void IService.StopService()
         {
             _status = ServiceStatus.Stopped;
+            if (FailedToStop)
+                throw new ArgumentException(nameof(FailedToStop));
         }
 
         // Set to true to cause the service to give
         // an error result when started
         public bool FailToStart { get; set; }
+
+        // Set to true to cause the service to give
+        // an error result when stopped
+        public bool FailedToStop { get; set; }
     }
 }
