+++ src/NUnitEngine/nunit.engine/Services/ServiceManager.cs
